Restructure parameterization and use a multistep method for history.
