+++ ksp_plugin/ksp_plugin.vcxproj.filters
@@ -56,6 +56,9 @@
     <ClInclude Include="identification.hpp">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="integrators.hpp">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="interface.cpp">
@@ -104,6 +107,9 @@
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="celestial.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="integrators.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
