+++ ksp_plugin_test/ksp_plugin_test.vcxproj.filters
@@ -101,6 +101,9 @@
     <ClCompile Include="..\ksp_plugin\celestial.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="..\ksp_plugin\integrators.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="mock_plugin.hpp">
